mqtt_to_postgres/requirements.txt
paho-mqtt
psycopg2
